requirements.txt
openai
backoff
python-dotenv